// src/components/game/hooks/heroCombat/types.ts
export interface UseHeroCombatActionsProps {
  entities: CombatEntity[];
  setEntities: React.Dispatch<React.SetStateAction<CombatEntity[]>>;
  activeEntity: CombatEntity | undefined;
  isHeroTurn: boolean;
  isBattleOver: boolean;
  character: any;
  currentSelectedAttack: any;
  characterAttacks: any[];
  selectedAttackIndex: number;
  weaponMasteryInfo: any;
  biome: BiomeType;
  weather: WeatherType;
  isNight: boolean;
  torches: GridPosition[];
  grid: CellData[][];
  activeAdvantageMode: 'normal' | 'advantage' | 'disadvantage';
  gwmActive: boolean;
  sharpshooterActive: boolean;
  activeLargeForm: boolean;
  isGoliath: boolean;
  goliathAncestryUses: number;
  isHalfling: boolean;
  hasHeroicInspiration: boolean;
  isOrc: boolean;
  adrenalineRushUses: number;
  adrenalineRushMaxUses: number;
  healingHandsUses: number;
  celestialRevelationUses: number;
  channelDivinityUses: number;
  channelDivinityMaxUses: number;
  breathWeaponUses: number;
  breathWeaponDetails: any;
  spellSlots: number;
  spellSlotsMax: number;
  secondWindUses: number;
  secondWindMaxUses: number;
  indomitableUses: number;
  indomitableMaxUses: number;
  recklessAttackActive: boolean;
  bardicInspirationUses: number;
  bardicInspirationMaxUses: number;
  layOnHandsPool: number;
  layOnHandsMaxPool: number;
  focusPointsUses: number;
  focusPointsMaxUses: number;
  wildShapeUses: number;
  wildShapeMaxUses: number;
  superiorityDiceUses: number;
  superiorityDiceMaxUses: number;
  pendingTacticalMindInfo: any;
  isSfxEnabled: boolean;

  // Setters & Actions
  addCombatLog: (speaker: string, title: string, detail: string, type?: any) => void;
  triggerAttackVisualEffect: (start: GridPosition, end: GridPosition, isRanged: boolean, hit: boolean, damage: number, isCritical: boolean) => void;
  setLatestRoll: (roll: any) => void;
  processDamageAndCheckKill: (targetId: string, rawDamage: number, attackerName: string, damageType: string, attackerId: string) => void;
  moveHeroDirection: (dx: number, dy: number) => void;
  consumeThrownWeapon: (name: string) => void;
  setDroppedLoot: React.Dispatch<React.SetStateAction<LootItem[]>>;
  checkAmmunitionRequirement: (atk: any) => any;
  getCharacterAmmoCount: (req: any) => number;
  consumeAmmunition: (req: any) => void;
  getActiveFeats: () => string[];
  getHeroLightRadiusInCells: () => number;
  isEntityVisible: (ent: CombatEntity) => boolean;
  setRollAdvantageState: (val: 'normal' | 'advantage' | 'disadvantage') => void;
  setPendingGoliathHitInfo: (info: any) => void;
  setTargetCandidates: (candidates: CombatEntity[]) => void;
  setPendingAttackInfo: (info: any) => void;
  setShowTargetModal: (val: boolean) => void;
  setShowAttackModal: (val: boolean) => void;
  setPendingHalflingLuckInfo: (info: any) => void;
  setPendingHeroicInspirationInfo: (info: any) => void;
  setFloatingTexts: React.Dispatch<React.SetStateAction<any[]>>;
  setAdrenalineRushUses: React.Dispatch<React.SetStateAction<number>>;
  setHealingHandsUses: React.Dispatch<React.SetStateAction<number>>;
  setCelestialRevelationUses: React.Dispatch<React.SetStateAction<number>>;
  setShowRevelationMenu: (val: boolean) => void;
  handleExecuteCelestialRevelation: (type: any) => void;
  setChannelDivinityUses: React.Dispatch<React.SetStateAction<number>>;
  setSelectedBreathTargets: (targets: CombatEntity[]) => void;
  setBreathWeaponShape: (shape: 'cone' | 'line') => void;
  setShowBreathWeaponModal: (val: boolean) => void;
  setBreathWeaponUses: React.Dispatch<React.SetStateAction<number>>;
  setActiveEffects: React.Dispatch<React.SetStateAction<any[]>>;
  setSpellSlots: React.Dispatch<React.SetStateAction<number>>;
  setSecondWindUses: React.Dispatch<React.SetStateAction<number>>;
  setPendingTacticalMindInfo: (info: any) => void;
  setShowTacticalMindAlertModal: (val: boolean) => void;
  setIndomitableUses: React.Dispatch<React.SetStateAction<number>>;
  setRecklessAttackActive: React.Dispatch<React.SetStateAction<boolean>>;
  setBardicInspirationUses: React.Dispatch<React.SetStateAction<number>>;
  setLayOnHandsPool: React.Dispatch<React.SetStateAction<number>>;
  setFocusPointsUses: React.Dispatch<React.SetStateAction<number>>;
  setWildShapeUses: React.Dispatch<React.SetStateAction<number>>;
  setSuperiorityDiceUses: React.Dispatch<React.SetStateAction<number>>;
}